backend/requirements.txt
redisvl
redis
langchain-google-genai
python-dotenv